moved novu to it's own stack
